requirements/base.txt
Django==1.9.4
wheel==0.24.0

psycopg2==2.6.1
requests==2.9.1
django-model-utils==2.4
django-markdown-deux
django-localflavor==1.2
django-extensions==1.6.1
django-redis

git+https://github.com/DemocracyClub/django-static-precompiler
git+git://github.com/DemocracyClub/dc_base_theme.git
